Separate unstructured and tensor norm functions.
With a unified `norms.hpp`, the tensor product code becomes (needlessly)
dependent on MOAB.
